Started on snippet toolbar button
